apps/ncsoccer-scraper/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ncsoccer-scraper"
version = "0.1.0"
description = "CLI tool to track Key West FC across NC Soccer Hudson leagues"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",
    "rich>=13.0.0",
    "click>=8.1.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "pyyaml>=6.0.0",
    "boto3>=1.34.0",
    "mcp[cli]>=1.0.0",
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "pgvector>=0.3.0",
    "openai>=1.0.0",
]

[project.scripts]
soccer = "ncsoccer.cli:cli"
ncsoccer-mcp = "ncsoccer.mcp_server:main"

[tool.hatch.build.targets.wheel]
packages = ["src/ncsoccer"]
